web/requirements.txt
asgiref==3.7.2
Django==5.0.3
psycopg==3.1.18
psycopg-binary==3.1.18
sqlparse==0.4.4
django-allauth==0.61.1
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
black==24.2.0
django-crispy-forms==2.1
crispy-bulma==0.11.0
coverage==7.4.4
black==24.2.0
django-environ==0.11.2
pillow==10.2.0
django-storages==1.14.2
boto3==1.34.69
django-debug-toolbar==4.3.0
django-extensions==3.2.3
ipython==8.22.2
gunicorn==21.2.0
